APIs/azure.com/applicationinsights-eaSubscriptionMigration_API/2017-10-01/python-aiohttp-client/tests/test_default_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.cloud_error import CloudError
from openapi_server.models.ea_subscription_migration_date import EASubscriptionMigrationDate


pytestmark = pytest.mark.asyncio

async def test_e_a_subscription_list_migration_date_post(client):
    """Test case for e_a_subscription_list_migration_date_post

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/subscriptions/{subscription_id}/providers/microsoft.insights/listMigrationdate'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_e_a_subscription_migrate_to_new_pricing_model_post(client):
    """Test case for e_a_subscription_migrate_to_new_pricing_model_post

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/subscriptions/{subscription_id}/providers/microsoft.insights/migrateToNewPricingModel'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_e_a_subscription_rollback_to_legacy_pricing_model_post(client):
    """Test case for e_a_subscription_rollback_to_legacy_pricing_model_post

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/subscriptions/{subscription_id}/providers/microsoft.insights/rollbackToLegacyPricingModel'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

